Fix the test program for rafl.
Since rafl has a CUDA dependency, test-rafl does as well. This is a particular issue on Mac OS X because UseCUDA.cmake makes the compilation of rafl use the -stdlib=libstdc++ flag. If test-rafl is then built against a different version of the C++ standard library, obscure linker errors result.